Separating footer embed JS code into its own file
