+++ readthedocs/core/static/core/js/readthedocs-doc-embed.js
@@ -1,1 +1,1 @@
-!function t(e,o,r){function n(a,s){if(!o[a]){if(!e[a]){var c="function"==typeof require&&require;if(!s&&c)return c(a,!0);if(i)return i(a,!0);var d=new Error("Cannot find module '"+a+"'");throw d.code="MODULE_NOT_FOUND",d}var p=o[a]={exports:{}};e[a][0].call(p.exports,function(t){var o=e[a][1][t];return n(o?o:t)},p,p.exports,t,e,o,r)}return o[a].exports}for(var i="function"==typeof require&&require,a=0;a<r.length;a++)n(r[a]);return n}({1:[function(t,e,o){function r(t){this.config=t,"sphinx_rtd_theme"!=this.config.theme&&1==$("div.rst-other-versions").length&&(this.config.theme="sphinx_rtd_theme"),void 0==this.config.api_host&&(this.config.api_host="https://readthedocs.org")}e.exports={Build:r},r.prototype.is_rtd_theme=function(){return"sphinx_rtd_theme"==this.config.theme},r.prototype.is_sphinx_builder=function(){return!("builder"in this.config)||"mkdocs"!=this.config.builder},r.prototype.show_promo=function(){return"https://readthedocs.com"!=this.config.api_host&&this.is_sphinx_builder()&&this.is_rtd_theme()}},{}],2:[function(t,e,o){var r=t("./sponsorship"),n=t("./doc");$(document).ready(function(){function t(t){var e=t.closest("li");e.siblings("li.current").removeClass("current"),e.siblings().find("li.current").removeClass("current"),e.find("> ul li.current").removeClass("current"),e.toggleClass("current")}function e(t){if(!t.is_highest){var e=window.location.pathname.replace(READTHEDOCS_DATA.version,t.slug),o=$('<div class="admonition warning"> <p class="first                          admonition-title">Note</p> <p class="last">                          You are not using the most up to date version                          of the library. <a href="#"></a> is the newest version.</p>                         </div>');o.find("a").attr("href",e).text(t.version),body=$("div.body"),body.length||(body=$("div.document")),body.prepend(o)}}var o=new n.Build(READTHEDOCS_DATA);get_data={project:READTHEDOCS_DATA.project,version:READTHEDOCS_DATA.version,page:READTHEDOCS_DATA.page,theme:READTHEDOCS_DATA.theme,format:"jsonp"},"docroot"in READTHEDOCS_DATA&&(get_data.docroot=READTHEDOCS_DATA.docroot),"source_suffix"in READTHEDOCS_DATA&&(get_data.source_suffix=READTHEDOCS_DATA.source_suffix);var i=READTHEDOCS_DATA.api_host;void 0===i&&(i="https://readthedocs.org"),0===window.location.pathname.indexOf("/projects/")&&(get_data.subproject=!0),$.ajax({url:i+"/api/v2/footer_html/",crossDomain:!0,xhrFields:{withCredentials:!0},dataType:"jsonp",data:get_data,success:function(t){function e(t){return/^(GET|HEAD|OPTIONS|TRACE)$/.test(t)}if(o.is_rtd_theme()?$("div.rst-other-versions").html(t.html):$("body").append(t.html),t.version_active?!t.version_supported:$(".rst-current-version").addClass("rst-out-of-date"),t.promo&&o.show_promo()){var n=new r.Promo(t.promo_data.id,t.promo_data.text,t.promo_data.link,t.promo_data.image);n&&n.display()}$.ajaxSetup({beforeSend:function(t,o){e(o.type)||t.setRequestHeader("X-CSRFToken",$("a.bookmark[token]").attr("token"))}})},error:function(){console.log("Error loading Read the Docs footer")}}),(!("builder"in READTHEDOCS_DATA)||"builder"in READTHEDOCS_DATA&&"mkdocs"!=READTHEDOCS_DATA.builder)&&($(document).on("click","[data-toggle='wy-nav-top']",function(){$("[data-toggle='wy-nav-shift']").toggleClass("shift"),$("[data-toggle='rst-versions']").toggleClass("shift")}),$(document).on("click",".wy-menu-vertical .current ul li a",function(){var e=$(this);$("[data-toggle='wy-nav-shift']").removeClass("shift"),$("[data-toggle='rst-versions']").toggleClass("shift"),t(e),"undefined"!=typeof window.SphinxRtdTheme&&window.SphinxRtdTheme.StickyNav.hashChange()}),$(document).on("click","[data-toggle='rst-current-version']",function(){$("[data-toggle='rst-versions']").toggleClass("shift-up")}),$("table.docutils:not(.field-list)").wrap("<div class='wy-table-responsive'></div>"),$(".wy-menu-vertical ul").siblings("a").each(function(){var e=$(this);expand=$('<span class="toctree-expand"></span>'),expand.on("click",function(o){return t(e),o.stopPropagation(),!1}),e.prepend(expand)}),window.SphinxRtdTheme=function(t){var e=function(){var e,o,r=!1,n=!1,i=0,a=function(){s(),c(),o.on("hashchange",c),o.on("scroll",function(){n||(r=!0)}),setInterval(function(){if(r){r=!1;var t=o.scrollTop(),n=e.scrollTop(),a=n+(t-i);e.scrollTop(a),i=t}},25)},s=function(){e=t("nav.wy-nav-side:first"),o=t(window)},c=function(){var t=encodeURI(window.location.hash);if(t)try{var e=$(".wy-menu-vertical").find('[href="'+t+'"]');$(".wy-menu-vertical li.toctree-l1 li.current").removeClass("current"),e.closest("li.toctree-l2").addClass("current"),e.closest("li.toctree-l3").addClass("current"),e.closest("li.toctree-l4").addClass("current")}catch(o){console.log("Error expanding nav for anchor",o)}},d=function(){n=!0,o.one("hashchange",function(){n=!1})};return t(s),{enable:a,hashChange:d}}();return{StickyNav:e}}($)),$.ajax({url:"https://api.grokthedocs.com/static/javascript/bundle-client.js",crossDomain:!0,dataType:"script"});var a=[i+"/api/v1/version/",READTHEDOCS_DATA.project,"/highest/",READTHEDOCS_DATA.version,"/?callback=?"].join("");if($.getJSON(a,e),"builder"in READTHEDOCS_DATA&&"mkdocs"==READTHEDOCS_DATA.builder){$("<input>").attr({type:"hidden",name:"project",value:READTHEDOCS_DATA.project}).appendTo("#rtd-search-form"),$("<input>").attr({type:"hidden",name:"version",value:READTHEDOCS_DATA.version}).appendTo("#rtd-search-form"),$("<input>").attr({type:"hidden",name:"type",value:"file"}).appendTo("#rtd-search-form"),$("#rtd-search-form").prop("action",i+"/elasticsearch/");var s=$("nav.wy-nav-side:first"),c=$(window),d="stickynav",p=function(){s.height()<=c.height()?s.addClass(d):s.removeClass(d)};c.on("resize",p),p()}})},{"./doc":1,"./sponsorship":3}],3:[function(t,e,o){function r(t,e,o,r){this.id=t,this.text=e,this.link=o,this.image=r,this.promo=null}var n=window.$;e.exports={Promo:r},r.prototype.create=function(){var t=this,e=n("nav.wy-nav-side");if(e.length){promo=n("<div />").attr("class","wy-menu rst-pro");{var o=n("<div />").attr("class","rst-pro-about"),r=n("<a />").attr("href","http://docs.readthedocs.org/en/latest/sponsors.html#sponsorship-information").appendTo(o);n("<i />").attr("class","fa fa-info-circle").appendTo(r)}if(o.appendTo(promo),t.image){{var i=n("<a />").attr("class","rst-pro-image-wrapper").attr("href",t.link);n("<img />").attr("class","rst-pro-image").attr("src",t.image).appendTo(i)}promo.append(i)}var a=n("<span />").html(t.text);return n(a).find("a").each(function(){n(this).attr("class","rst-pro-link").attr("href",t.link).attr("target","_blank").on("click",function(e){_gaq&&_gaq.push(["rtfd._setAccount","UA-17997319-1"],["rtfd._trackEvent","Promo","Click",t.id])})}),promo.append(a),promo.appendTo(e),promo.wrapper=n("<div />").attr("class","rst-pro-wrapper").appendTo(e),promo}},r.prototype.display=function(){var t=this.promo;t||(t=this.promo=this.create()),t.show()},r.prototype.disable=function(){},r.from_variants=function(t){if(0==t.length)return null;var e=Math.floor(Math.random()*t.length),o=t[e],n=o.text,i=o.link,a=o.image,s=o.id;return new r(s,n,i,a)}},{}]},{},[2]);
+!function t(e,o,r){function n(a,s){if(!o[a]){if(!e[a]){var c="function"==typeof require&&require;if(!s&&c)return c(a,!0);if(i)return i(a,!0);var d=new Error("Cannot find module '"+a+"'");throw d.code="MODULE_NOT_FOUND",d}var l=o[a]={exports:{}};e[a][0].call(l.exports,function(t){var o=e[a][1][t];return n(o?o:t)},l,l.exports,t,e,o,r)}return o[a].exports}for(var i="function"==typeof require&&require,a=0;a<r.length;a++)n(r[a]);return n}({1:[function(t,e,o){function r(t){var e={project:n.project,version:n.version,page:n.page,theme:n.theme,format:"jsonp"};"docroot"in n&&(e.docroot=n.docroot),"source_suffix"in n&&(e.source_suffix=n.source_suffix),0===window.location.pathname.indexOf("/projects/")&&(e.subproject=!0),$.ajax({url:n.api_host+"/api/v2/footer_html/",crossDomain:!0,xhrFields:{withCredentials:!0},dataType:"jsonp",data:e,success:function(e){function o(t){return/^(GET|HEAD|OPTIONS|TRACE)$/.test(t)}if(t.is_rtd_theme()?$("div.rst-other-versions").html(e.html):$("body").append(e.html),e.version_active?!e.version_supported:$(".rst-current-version").addClass("rst-out-of-date"),e.promo&&t.show_promo()){var r=new sponsorship.Promo(e.promo_data.id,e.promo_data.text,e.promo_data.link,e.promo_data.image);r&&r.display()}$.ajaxSetup({beforeSend:function(t,e){o(e.type)||t.setRequestHeader("X-CSRFToken",$("a.bookmark[token]").attr("token"))}})},error:function(){console.log("Error loading Read the Docs footer")}})}var n=t("./rtd-data");e.exports={init:r}},{"./rtd-data":3}],2:[function(t,e,o){function r(){$.ajax({url:"https://api.grokthedocs.com/static/javascript/bundle-client.js",crossDomain:!0,dataType:"script"})}e.exports={init:r}},{}],3:[function(t,e,o){var r=READTHEDOCS_DATA;void 0===r.api_host&&(r.api_host="https://readthedocs.org"),e.exports=r},{}],4:[function(t,e,o){function r(t){this.config=t,"sphinx_rtd_theme"!=this.config.theme&&1==$("div.rst-other-versions").length&&(this.config.theme="sphinx_rtd_theme"),void 0==this.config.api_host&&(this.config.api_host="https://readthedocs.org")}e.exports={Build:r},r.prototype.is_rtd_theme=function(){return"sphinx_rtd_theme"==this.config.theme},r.prototype.is_sphinx_builder=function(){return!("builder"in this.config)||"mkdocs"!=this.config.builder},r.prototype.show_promo=function(){return"https://readthedocs.com"!=this.config.api_host&&this.is_sphinx_builder()&&this.is_rtd_theme()}},{}],5:[function(t,e,o){var r=(t("./sponsorship"),t("./doc")),n=t("./doc-embed/footer.js"),i=t("./doc-embed/grokthedocs-client"),a=t("./doc-embed/rtd-data");$(document).ready(function(){function t(t){var e=t.closest("li");e.siblings("li.current").removeClass("current"),e.siblings().find("li.current").removeClass("current"),e.find("> ul li.current").removeClass("current"),e.toggleClass("current")}function e(t){if(!t.is_highest){var e=window.location.pathname.replace(a.version,t.slug),o=$('<div class="admonition warning"> <p class="first                          admonition-title">Note</p> <p class="last">                          You are not using the most up to date version                          of the library. <a href="#"></a> is the newest version.</p>                         </div>');o.find("a").attr("href",e).text(t.version),body=$("div.body"),body.length||(body=$("div.document")),body.prepend(o)}}function o(){var t=$.getQueryParameters(),e=t.q?t.q[0].split(/\s+/):[],o=!0;e.length&&(e=e.join(" "),console.log("Searching based on GET arg for: "+e),$("#rtd-search-form input[name='q']").val(e),s(e,o))}function s(t,e){var o={project:a.project,version:a.version,format:"jsonp",q:t};$.ajax({url:a.api_host+"/api/v2/search/section/",crossDomain:!0,xhrFields:{withCredentials:!0},dataType:"jsonp",data:o,success:function(o){p(e),hits=o.results.hits.hits,hits.length?d(hits,t):l()},error:function(){console.log("Error searching")}})}function c(t){var e=$(".rst-content");e.html(t)}function d(t,e){FIRSTRUN={},current=$(".toctree-l1.current > a");for(var o in t){var r=t[o],n=r.fields.path,i=r.fields.page_id,a=r.fields.title,s=r.fields.content,c=(r.highlight.content,r._score),d=$(".toctree-l1 > a[href^='"+n+"']"),l=d.next();console.log(n),FIRSTRUN[n]||(d.show(),d.attr("href",d.attr("href")+"?highlight="+e),d.parent().addClass("current"),d.append("<i style='position:absolute;right:30px;top:6px;' class='fa fa-search result-icon'></i>"),l.empty(),FIRSTRUN[n]=!0),FIRSTRUN[n+a]||(l.append('<li class="toctree-l2"><a class="reference internal search-result" pageId="'+i+'">'+a+'</a><span style="display: none;" class="data">'+s+"</span></li>"),c>1&&($(".toctree-l2 "),inserted=$('.toctree-l2 > [pageId="'+i+'"]'),inserted.append("<i style='position:absolute;right:30px;top:6px;' class='fa fa-fire'></i>")),FIRSTRUN[n+a]=!0)}$.each($(".toctree-l1 > a"),function(t,e){hide=!0,""===$(e).attr("href")&&(hide=!1);for(var o in FIRSTRUN)0===$(e).attr("href").indexOf(o)&&(hide=!1);hide&&$(e).hide()})}function l(){$.each($(".toctree-l1 > a"),function(t,e){var e=$(e);e.show(),e.parent().show()})}function p(t){$(".result-icon").remove(),$.each($(".toctree-l1 > a"),function(e,o){var o=$(o);t&&(o.parent().removeClass("current"),o.next().empty())})}var h=new r.Build(a);n.init(h),(!("builder"in a)||"builder"in a&&"mkdocs"!=a.builder)&&($(document).on("click","[data-toggle='wy-nav-top']",function(){$("[data-toggle='wy-nav-shift']").toggleClass("shift"),$("[data-toggle='rst-versions']").toggleClass("shift")}),$(document).on("click",".wy-menu-vertical .current ul li a",function(){var e=$(this);$("[data-toggle='wy-nav-shift']").removeClass("shift"),$("[data-toggle='rst-versions']").toggleClass("shift"),t(e),"undefined"!=typeof window.SphinxRtdTheme&&window.SphinxRtdTheme.StickyNav.hashChange()}),$(document).on("click","[data-toggle='rst-current-version']",function(){$("[data-toggle='rst-versions']").toggleClass("shift-up")}),$("table.docutils:not(.field-list)").wrap("<div class='wy-table-responsive'></div>"),$(".wy-menu-vertical ul").siblings("a").each(function(){var e=$(this);expand=$('<span class="toctree-expand"></span>'),expand.on("click",function(o){return t(e),o.stopPropagation(),!1}),e.prepend(expand)}),window.SphinxRtdTheme=function(t){var e=function(){var e,o,r=!1,n=!1,i=0,a=function(){s(),c(),o.on("hashchange",c),o.on("scroll",function(){n||(r=!0)}),setInterval(function(){if(r){r=!1;var t=o.scrollTop(),n=e.scrollTop(),a=n+(t-i);e.scrollTop(a),i=t}},25)},s=function(){e=t("nav.wy-nav-side:first"),o=t(window)},c=function(){var t=encodeURI(window.location.hash);if(t)try{var e=$(".wy-menu-vertical").find('[href="'+t+'"]');$(".wy-menu-vertical li.toctree-l1 li.current").removeClass("current"),e.closest("li.toctree-l2").addClass("current"),e.closest("li.toctree-l3").addClass("current"),e.closest("li.toctree-l4").addClass("current")}catch(o){console.log("Error expanding nav for anchor",o)}},d=function(){n=!0,o.one("hashchange",function(){n=!1})};return t(s),{enable:a,hashChange:d}}();return{StickyNav:e}}($)),i.init();var u=[a.api_host+"/api/v1/version/",a.project,"/highest/",a.version,"/?callback=?"].join("");if($.getJSON(u,e),"builder"in a&&"mkdocs"==a.builder){$("<input>").attr({type:"hidden",name:"project",value:a.project}).appendTo("#rtd-search-form"),$("<input>").attr({type:"hidden",name:"version",value:a.version}).appendTo("#rtd-search-form"),$("<input>").attr({type:"hidden",name:"type",value:"file"}).appendTo("#rtd-search-form"),$("#rtd-search-form").prop("action",a.api_host+"/elasticsearch/");var f=$("nav.wy-nav-side:first"),m=$(window),v="stickynav",g=function(){f.height()<=m.height()?f.addClass(v):f.removeClass(v)};m.on("resize",g),g()}h.is_rtd_theme()&&o(),$(document).on({mouseenter:function(t){var e=$(t.target).next();e.show()},mouseleave:function(t){var e=$(t.target).next();e.hide()}},".result-count"),$(document).on("submit","#rtd-search-form",function(t){p();var e=$("#rtd-search-form input[name='q']").val();s(e,!0)}),$(document).on("click",".search-result",function(t){t.preventDefault(),html=$(t.target).next().html(),c(html)})})},{"./doc":4,"./doc-embed/footer.js":1,"./doc-embed/grokthedocs-client":2,"./doc-embed/rtd-data":3,"./sponsorship":6}],6:[function(t,e,o){function r(t,e,o,r){this.id=t,this.text=e,this.link=o,this.image=r,this.promo=null}var n=window.$;e.exports={Promo:r},r.prototype.create=function(){var t=this,e=n("nav.wy-nav-side");if(e.length){promo=n("<div />").attr("class","wy-menu rst-pro");{var o=n("<div />").attr("class","rst-pro-about"),r=n("<a />").attr("href","http://docs.readthedocs.org/en/latest/sponsors.html#sponsorship-information").appendTo(o);n("<i />").attr("class","fa fa-info-circle").appendTo(r)}if(o.appendTo(promo),t.image){{var i=n("<a />").attr("class","rst-pro-image-wrapper").attr("href",t.link);n("<img />").attr("class","rst-pro-image").attr("src",t.image).appendTo(i)}promo.append(i)}var a=n("<span />").html(t.text);return n(a).find("a").each(function(){n(this).attr("class","rst-pro-link").attr("href",t.link).attr("target","_blank").on("click",function(e){_gaq&&_gaq.push(["rtfd._setAccount","UA-17997319-1"],["rtfd._trackEvent","Promo","Click",t.id])})}),promo.append(a),promo.appendTo(e),promo.wrapper=n("<div />").attr("class","rst-pro-wrapper").appendTo(e),promo}},r.prototype.display=function(){var t=this.promo;t||(t=this.promo=this.create()),t.show()},r.prototype.disable=function(){},r.from_variants=function(t){if(0==t.length)return null;var e=Math.floor(Math.random()*t.length),o=t[e],n=o.text,i=o.link,a=o.image,s=o.id;return new r(s,n,i,a)}},{}]},{},[5]);
